Work around intermittent Summit scratch FS errors
In the last month or so, I've had several job chains die with I/O errors
when trying to restart from the latest checkpoint file. All the files
show up fine with `ls`, but reading from them fails for up to ~5-10
minutes. This PR works around this by restarting from the previous
checkpoint file if any of the files can't be read.